refactor: rename stream-listener to emit-listener
separate test code by feature
