+++ src/emiter/emit-observer.ts
@@ -12,7 +12,6 @@ export class EmitObserver<T> extends EventEmitter implements EmitObserverImpl<T>
 
   error(err: any): void {
     this.emit('error', err);
-    this.removeAllListeners();
   }
 
   complete(): void {
