+++ src/listen/emit-listener.ts
@@ -1,0 +1,3 @@
+export interface EmitListener {
+  unlisten: () => void;
+}
