--- src/listen/stream-listener.interface.ts
@@ -1,3 +1,0 @@
-export interface StreamListener {
-  unlisten: () => void;
-}
